Update - Reverted previous update due to unforseen negative consequences

+++ packages/scripts/tur-space/tur-space.sh
@@ -15,7 +15,7 @@ if [ -z "$config" ]; then
 fi
 
 if [ ! -e "$config" ]; then
-  echo "Error. The configuration can not be read from $config - Quitting."
+  echo "Error. The configuration can not be read from $config - Quitting." 
   echo "Make sure you specify its location in tur-space.sh"
   exit 1
 fi
@@ -127,7 +127,7 @@ proc_find_oldest_dir() {
       proc_debug "Warning: Did not find any dir to move/delete in $source_dir ($source_name). Skipping."
       proc_log "Warning: Did not find any dir to move/delete in $source_dir ($source_name). Skipping."
       if [ "$last_source_dir" = "$source_dir" ]; then
-        ## If no other incoming sections went ok, quit here or we'll get an endless loop.
+        ## If no other incoming sections went ok, quit here or we'll get an endless loop.        
         proc_debug "Error. No dirs found at all to process in any incoming section. Quitting."
         proc_log "Error. No dirs found at all to process in any incoming section. Quitting."
         exit 1
@@ -219,7 +219,7 @@ proc_find_destination_old() {
 
       while [ "$source_release_size" -ge "$free_space" ]; do
 
-proc_log "Starting loop. source_release_size: $source_release_size - free_space: $free_space - $destinations"
+proc_log "Starting loop. source_release_size: $source_release_size - free_space: $free_space - $destinations" 
 
         ## Make sure we dont del more then $MAX_LOOP things.
         loop_count=$[$loop_count+1]
@@ -237,7 +237,7 @@ proc_log "Starting loop. source_release_size: $source_release_size - free_space:
 
         if [ "$DEBUG" = "TRUE" ]; then IGNORES="$IGNORES|::$source_release::"; fi
 
-        MODE="ARCHIVE"
+        MODE="ARCHIVE"        
 proc_log  "Running proc_find_oldest_dir $source_release_path $source_name $source_device"
         proc_find_oldest_dir $source_release_path $source_name $source_device
 proc_log "proc_find_oldest_dir returned $oldest_dir_raw"
@@ -361,33 +361,21 @@ proc_move() {
       rm -rf "${dest_release_path}/${source_release}"
     fi
 
-    rsync_retries=0
-    while true; do
-        rsync $RSYNC_OPTIONS "${source_release_path}/${source_release}" "${dest_release_path}" && break
-        if [[ $rsync_retries -eq $MAX_RSYNC_RETRIES ]]; then
-            proc_log "Error: Rsync retries exceeded while copying $source_release from $source_release_path into $dest_release_path."
-            proc_debug "Error: Rsync retries exceeded while copying $source_release from $source_release_path into $dest_release_path."
-            proc_debug "Quitting and you should check why this happened on a simple copy."
-            sleep 5
-            proc_exit
-        fi
-        rsync_retries=$((rsync_retries + 1))
-    done
+    cp $COPY_OPTIONS "${source_release_path}/${source_release}" "${dest_release_path}"
 
-    ## Verification no longer required when using rsync
-    #num_files_source="`ls -1 "$source_release_path/$source_release" | wc -l | tr -d ' '`"
-    #num_files_dest="`ls -1 "$dest_release_path/$source_release" | wc -l | tr -d ' '`"
-    #if [ "$num_files_source" != "$num_files_dest" ]; then
-    #  sleep 5
-    #  num_files_source="`ls -1 "$source_release_path/$source_release" | wc -l | tr -d ' '`"
-    #  num_files_dest="`ls -1 "$dest_release_path/$source_release" | wc -l | tr -d ' '`"
-    #  if [ "$num_files_source" != "$num_files_dest" ]; then
-    #    proc_log "Error: After copying $source_release from $source_release_path into $dest_release_path, the number of files does not match up. Source:$num_files_source Dest:$num_files_dest"
-    #    proc_debug "Error. After copying $source_release from $source_release_path into $dest_release_path, the number of files does not match up. Source:$num_files_source Dest:$num_files_dest"
-    #    proc_debug "Quitting and you should check why this happened on a simple copy."
-    #    proc_exit
-    #  fi
-    #else
+    num_files_source="`ls -1 "$source_release_path/$source_release" | wc -l | tr -d ' '`"
+    num_files_dest="`ls -1 "$dest_release_path/$source_release" | wc -l | tr -d ' '`"
+    if [ "$num_files_source" != "$num_files_dest" ]; then
+      sleep 5
+      num_files_source="`ls -1 "$source_release_path/$source_release" | wc -l | tr -d ' '`"
+      num_files_dest="`ls -1 "$dest_release_path/$source_release" | wc -l | tr -d ' '`"
+      if [ "$num_files_source" != "$num_files_dest" ]; then
+        proc_log "Error: After copying $source_release from $source_release_path into $dest_release_path, the number of files does not match up. Source:$num_files_source Dest:$num_files_dest"
+        proc_debug "Error. After copying $source_release from $source_release_path into $dest_release_path, the number of files does not match up. Source:$num_files_source Dest:$num_files_dest"
+        proc_debug "Quitting and you should check why this happened on a simple copy."
+        proc_exit
+      fi
+    else
 
       ## Copy verified. Remove releases from original location.
       rm -rf "$source_release_path/$source_release"
@@ -398,7 +386,7 @@ proc_move() {
         chmod $CHMOD_OPTIONS "$dest_release_path/$source_release"
       fi
 
-    #fi
+    fi
   fi
 
   if [ "$DEBUG" = "TRUE" ]; then IGNORES="$IGNORES|::$source_release::"; fi
@@ -492,7 +480,7 @@ proc_find_destination() {
     proc_find_most_space_free $destinations
   fi
 }
-
+  
 proc_find_sourcedirs() {
   # trigger_device_check="$1"
 
@@ -524,7 +512,7 @@ proc_find_sourcedirs() {
     proc_debug "Oldest dir on $trigger_device_check is $source_name -> $source_release_path/$source_release."
 
     proc_find_destination $source_release_path $source_release $source_name
-  fi
+  fi    
 }
 
 
@@ -548,16 +536,11 @@ if [ -z "$TULS" ]; then
   TULS="tuls"
 fi
 
-RSYNC_OPTIONS="`grep "^RSYNC_OPTIONS=" $config | cut -d '=' -f2- | tr -d '"'`"
-if [ -z "$RSYNC_OPTIONS" ]; then
-  proc_log "Warning: RSYNC_OPTIONS not defined. Forcing '-ra'."
-  proc_debug "Warning: RSYNC_OPTIONS not defined. Forcing '-ra'."
-  RSYNC_OPTIONS="-ra"
-fi
-
-MAX_RSYNC_RETRIES="`grep "^MAX_RSYNC_RETRIES=" $config | cut -d '=' -f2 | tr -d '"'`"
-if [ -z "$MAX_RSYNC_RETRIES" ]; then
-  MAX_RSYNC_RETRIES="10"
+COPY_OPTIONS="`grep "^COPY_OPTIONS=" $config | cut -d '=' -f2- | tr -d '"'`"
+if [ -z "$COPY_OPTIONS" ]; then
+  proc_log "Warning: COPY_OPTIONS not defined. Forcing '-rf'."
+  proc_debug "Warning: COPY_OPTIONS not defined. Forcing '-rf'."
+  COPY_OPTIONS="-rf"
 fi
 
 CHOWN_OPTIONS="`grep "^CHOWN_OPTIONS=" $config | cut -d '=' -f2 | tr -d '"'`"
@@ -633,10 +616,9 @@ proc_sanity_check() {
 
   echo "Configuration file used: $config"
 
-  if [ "$RSYNC_OPTIONS" ]; then
-    echo "Rsync mode used: rsync $RSYNC_OPTIONS"
+  if [ "$COPY_OPTIONS" ]; then
+    echo "Copy mode used: cp $COPY_OPTIONS"
   fi
-  echo "A maximum of $MAX_RSYNC_RETRIES rsync retries will be attempted."
 
   if [ "$CHOWN_OPTIONS" ]; then
     echo "Moved releases owner will be set to $CHOWN_OPTIONS"
@@ -779,7 +761,7 @@ proc_sanity_check() {
   done
 
   proc_exit
-}
+} 
 
 proc_help() {
   echo ""
@@ -867,7 +849,7 @@ case $2 in
 esac
 
 if [ -e "$TMP/tur-space.lock" ]; then
-  if [ "$1" = "sanity" ]; then
+  if [ "$1" = "sanity" ]; then  
     DEBUG="TRUE"
   fi
